+++ ascii/const.go
@@ -4,8 +4,8 @@ const (
 	keyEscape                    = 27
 	ascii_palette                = "   ...',;:clodxkO0KXNWM"
 	ascii_palette_length         = len(ascii_palette) - 1
-	threshold_low        float32 = 0.1
-	threshold_high       float32 = 1 - threshold_low
+	threshold_low        float64 = 0.1
+	threshold_high       float64 = 1 - threshold_low
 )
 
 var (
